test: consolidate worker sessions HTTP characterization
