+++ pkg/services/worker_sessions/transports/http/handler_test.go
@@ -13,7 +13,6 @@ import (
 	"github.com/portpowered/infinite-you/pkg/services/providers"
 	"github.com/portpowered/infinite-you/pkg/services/work"
 	workersessions "github.com/portpowered/infinite-you/pkg/services/worker_sessions"
-	"github.com/portpowered/infinite-you/pkg/services/workers"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	"go.uber.org/zap"
 )
@@ -305,377 +304,6 @@ func TestListWorkerSessionsProjectsTopLevelScopeFiltersAndPagination(t *testing.
 	}
 	if response.PaginationContext == nil || response.PaginationContext.MaxResults != 1 || response.PaginationContext.NextToken == nil || *response.PaginationContext.NextToken != "Y3Vyc29yLTI=" {
 		t.Fatalf("pagination = %#v, want bounded continuation context", response.PaginationContext)
-	}
-}
-
-func TestGetWorkerSessionObservationBySessionIDProjectsFailureDiagnostics(t *testing.T) {
-	total := 17
-	duration := int64(2500)
-	failure := &workersessions.FailureCause{
-		Kind:                 workersessions.FailureCauseIncompleteOutput,
-		Detail:               "the Workers result did not include the required final output",
-		AgentRunFailureClass: workers.AgentRunFailureClassProvider,
-		ProviderFailureKind:  providers.ExecuteFailureKindDependency,
-	}
-	service := &fakeObservationService{getResult: workersessions.Observation{
-		WorkerSessionID:          "worker-session-1",
-		Model:                    stringPtr("gpt-5.6-luna"),
-		ReasoningEffort:          stringPtr("medium"),
-		ProviderSession:          providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-1"},
-		ProviderSessionAvailable: true,
-		WorkIDs:                  []string{"work-1"}, TurnID: "turn-1", AttemptID: "attempt-1",
-		State: workersessions.StateFailed, Duration: durationPtr(2500 * time.Millisecond),
-		DurationBasis: workersessions.DurationBasisRecordedTimestamps,
-		TokenUsage:    &workersessions.TokenUsage{TotalTokens: &total}, Transcript: workersessions.TranscriptAvailabilityAvailable,
-		TurnUsage: &workersessions.TurnUsage{TurnCount: 3, FinalContextTokens: 450, PeakContextTokens: 450},
-		Failure:   failure,
-		Parse:     workersessions.ParseDiagnostics{EventCount: 4, Errors: []workersessions.ParseDiagnostic{{Code: "provider_session_parse_error", LineNumber: 3, Message: "malformed event"}}},
-	}}
-	handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-	recorder := httptest.NewRecorder()
-	request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions/detail?provider=codex&kind=session_id&id=provider-session-1", nil)
-
-	handler.GetWorkerSessionObservationBySessionId(recorder, request, factoryapi.SessionID("session-1"), factoryapi.GetWorkerSessionObservationBySessionIdParams{
-		Provider: factoryapi.LoadableProviderSessionProvider("codex"), Kind: factoryapi.LoadableProviderSessionKind("session_id"), Id: "provider-session-1",
-	})
-
-	if recorder.Code != http.StatusOK {
-		t.Fatalf("status = %d, want 200; body=%s", recorder.Code, recorder.Body.String())
-	}
-	assertFailureObservationResponse(t, recorder.Body.Bytes(), service, total, duration)
-}
-
-func assertFailureObservationResponse(t *testing.T, payload []byte, service *fakeObservationService, total int, duration int64) {
-	t.Helper()
-	var response factoryapi.WorkerSessionObservation
-	if err := json.Unmarshal(payload, &response); err != nil {
-		t.Fatalf("decode response: %v", err)
-	}
-	assertFailureObservationIdentity(t, response)
-	assertFailureObservationCause(t, response)
-	assertFailureObservationUsage(t, response, total, duration)
-	assertFailureObservationTurnUsage(t, response)
-	assertFailureObservationParse(t, response)
-	assertFailureObservationExecutionFacts(t, response)
-	assertFailureObservationRequest(t, service)
-}
-
-func assertFailureObservationIdentity(t *testing.T, response factoryapi.WorkerSessionObservation) {
-	t.Helper()
-	if response.WorkerSessionId != "worker-session-1" || response.State != factoryapi.WorkerSessionObservationStateFailed || response.AttemptId != "attempt-1" {
-		t.Fatalf("identity/state = %#v, want failed attempt projection", response)
-	}
-}
-
-func assertFailureObservationCause(t *testing.T, response factoryapi.WorkerSessionObservation) {
-	t.Helper()
-	if response.Failure == nil || response.Failure.Kind != string(workersessions.FailureCauseIncompleteOutput) ||
-		response.Failure.Detail != "the Workers result did not include the required final output" || response.Failure.ProviderFailureKind == nil ||
-		response.Failure.AgentRunFailureClass == nil || *response.Failure.AgentRunFailureClass != workers.AgentRunFailureClassProvider {
-		t.Fatalf("failure = %#v, want structured failure diagnostics", response.Failure)
-	}
-}
-
-func assertFailureObservationUsage(t *testing.T, response factoryapi.WorkerSessionObservation, total int, duration int64) {
-	t.Helper()
-	if response.TokenUsage == nil || response.TokenUsage.TotalTokens == nil || *response.TokenUsage.TotalTokens != total || response.DurationMillis == nil || *response.DurationMillis != duration {
-		t.Fatalf("usage/duration = %#v/%v, want %d/%d", response.TokenUsage, response.DurationMillis, total, duration)
-	}
-}
-
-func assertFailureObservationParse(t *testing.T, response factoryapi.WorkerSessionObservation) {
-	t.Helper()
-	if response.Parse.EventCount != 4 || len(response.Parse.Errors) != 1 {
-		t.Fatalf("parse = %#v, want event and parse diagnostics", response.Parse)
-	}
-}
-
-func assertFailureObservationRequest(t *testing.T, service *fakeObservationService) {
-	t.Helper()
-	if service.getProviderSession != (providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-1"}) {
-		t.Fatalf("service identity = %#v, want exact provider session ref", service.getProviderSession)
-	}
-}
-
-func TestReadWorkerSessionTranscriptBySessionIDProjectsNormalizedEntries(t *testing.T) {
-	text := "assistant response"
-	toolName := "lookup"
-	arguments := `{"key":"value"}`
-	service := &fakeObservationService{readResult: workersessions.ReadTranscriptResult{
-		WorkerSessionID: "worker-session-1",
-		ProviderSession: providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-1"},
-		WorkIDs:         []string{"work-1"}, TurnID: "turn-1", AttemptID: "attempt-1", State: workersessions.StateFailed,
-		Entries: []workersessions.TranscriptEntry{
-			{Order: 1, Type: workersessions.TranscriptToolCall, Name: &toolName, Arguments: &arguments},
-			{Order: 2, Type: workersessions.TranscriptAssistantMessage, Text: &text},
-		},
-	}}
-	handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-	recorder := httptest.NewRecorder()
-	request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions/transcript", nil)
-	handler.ReadWorkerSessionTranscriptBySessionId(recorder, request, factoryapi.SessionID("session-1"), factoryapi.ReadWorkerSessionTranscriptBySessionIdParams{
-		Provider: factoryapi.LoadableProviderSessionProvider("codex"), Kind: factoryapi.LoadableProviderSessionKind("session_id"), Id: "provider-session-1",
-	})
-
-	if recorder.Code != http.StatusOK {
-		t.Fatalf("status = %d, want 200; body=%s", recorder.Code, recorder.Body.String())
-	}
-	assertTranscriptResponse(t, recorder.Body.Bytes(), service, text, toolName)
-}
-
-func assertTranscriptResponse(t *testing.T, payload []byte, service *fakeObservationService, text, toolName string) {
-	t.Helper()
-	var response factoryapi.WorkerSessionTranscriptResponse
-	if err := json.Unmarshal(payload, &response); err != nil {
-		t.Fatalf("decode response: %v", err)
-	}
-	if response.WorkerSessionId != "worker-session-1" || response.State != string(workersessions.StateFailed) || response.AttemptId != "attempt-1" || response.TurnId == nil || *response.TurnId != "turn-1" {
-		t.Fatalf("response envelope = %#v, want correlated terminal session", response)
-	}
-	if len(response.Entries) != 2 || response.Entries[0].Type != factoryapi.ProviderSessionTranscriptEntryType(workersessions.TranscriptToolCall) || response.Entries[0].Name == nil || *response.Entries[0].Name != toolName || response.Entries[1].Text == nil || *response.Entries[1].Text != text {
-		t.Fatalf("entries = %#v, want ordered normalized tool and assistant entries", response.Entries)
-	}
-	wantRef := providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-1"}
-	if !service.readCalled || service.readProviderSession != wantRef {
-		t.Fatalf("read request = called=%t ref=%#v, want exact %v", service.readCalled, service.readProviderSession, wantRef)
-	}
-}
-
-func TestReadWorkerSessionTranscriptBySessionIDMapsDistinctTranscriptFailures(t *testing.T) {
-	for _, testCase := range []struct {
-		name   string
-		err    error
-		status int
-		code   factoryapi.ErrorResponseCode
-	}{
-		{name: "missing", err: workersessions.ErrObservationSessionNotFound, status: http.StatusNotFound, code: factoryapi.ErrorResponseCodeNOTFOUND},
-		{name: "active", err: workersessions.ErrObservationTranscriptActive, status: http.StatusConflict, code: factoryapi.ErrorResponseCodeWORKERSESSIONTRANSCRIPTACTIVE},
-		{name: "unavailable", err: workersessions.ErrObservationTranscriptUnavailable, status: http.StatusInternalServerError, code: factoryapi.ErrorResponseCodeWORKERSESSIONTRANSCRIPTUNAVAILABLE},
-		{name: "projection", err: workersessions.ErrObservationTranscriptProjectionUnavailable, status: http.StatusInternalServerError, code: factoryapi.ErrorResponseCodeWORKERSESSIONTRANSCRIPTPROJECTIONUNAVAILABLE},
-	} {
-		t.Run(testCase.name, func(t *testing.T) {
-			service := &fakeObservationService{readErr: testCase.err}
-			handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-			recorder := httptest.NewRecorder()
-			request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions/transcript", nil)
-			handler.ReadWorkerSessionTranscriptBySessionId(recorder, request, factoryapi.SessionID("session-1"), factoryapi.ReadWorkerSessionTranscriptBySessionIdParams{
-				Provider: factoryapi.LoadableProviderSessionProvider("codex"), Kind: factoryapi.LoadableProviderSessionKind("session_id"), Id: "provider-session-1",
-			})
-			if recorder.Code != testCase.status {
-				t.Fatalf("status = %d, want %d; body=%s", recorder.Code, testCase.status, recorder.Body.String())
-			}
-			var response factoryapi.ErrorResponse
-			if err := json.Unmarshal(recorder.Body.Bytes(), &response); err != nil {
-				t.Fatalf("decode error response: %v", err)
-			}
-			if response.Code != testCase.code {
-				t.Fatalf("error code = %q, want %q", response.Code, testCase.code)
-			}
-		})
-	}
-}
-
-func TestAdapterGetWorkerSessionObservationUsesExactProviderSessionIdentity(t *testing.T) {
-	service := &fakeObservationService{getResult: workersessions.Observation{
-		WorkerSessionID: "worker-session-1", ProviderSessionAvailable: true,
-		ProviderSession: providers.SessionRef{Provider: providers.IDCursor, Kind: providers.SessionIDKind, ID: "cursor-session-1"},
-		AttemptID:       "attempt-1", State: workersessions.StateCompleted,
-		DurationBasis: workersessions.DurationBasisUnavailable, Transcript: workersessions.TranscriptAvailabilityUnavailable,
-	}}
-	adapter := NewAdapter(service, workServiceStub{})
-	response, err := adapter.GetWorkerSessionObservation(context.Background(), "session-1", " cursor ", " session_id ", " cursor-session-1 ")
-	if err != nil {
-		t.Fatalf("GetWorkerSessionObservation() error = %v", err)
-	}
-	if response.WorkerSessionId != "worker-session-1" || response.ProviderSession == nil || response.ProviderSession.Id != "cursor-session-1" {
-		t.Fatalf("response = %#v, want detached observation", response)
-	}
-	want := providers.SessionRef{Provider: providers.IDCursor, Kind: providers.SessionIDKind, ID: "cursor-session-1"}
-	if service.getProviderSession != want {
-		t.Fatalf("service identity = %#v, want %#v", service.getProviderSession, want)
-	}
-}
-
-func TestGetWorkerSessionObservationBySessionIDMapsMissingAndUnavailable(t *testing.T) {
-	for _, testCase := range []struct {
-		name   string
-		err    error
-		status int
-		code   factoryapi.ErrorResponseCode
-	}{
-		{name: "missing", err: workersessions.ErrObservationSessionNotFound, status: http.StatusNotFound, code: factoryapi.ErrorResponseCodeNOTFOUND},
-		{name: "unavailable", err: workersessions.ErrObservationProjectionUnavailable, status: http.StatusInternalServerError, code: factoryapi.ErrorResponseCode("PROJECTION_UNAVAILABLE")},
-	} {
-		t.Run(testCase.name, func(t *testing.T) {
-			service := &fakeObservationService{getErr: testCase.err}
-			handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-			recorder := httptest.NewRecorder()
-			request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions/detail", nil)
-			handler.GetWorkerSessionObservationBySessionId(recorder, request, factoryapi.SessionID("session-1"), factoryapi.GetWorkerSessionObservationBySessionIdParams{
-				Provider: factoryapi.LoadableProviderSessionProvider("codex"), Kind: factoryapi.LoadableProviderSessionKind("session_id"), Id: "missing",
-			})
-			if recorder.Code != testCase.status {
-				t.Fatalf("status = %d, want %d; body=%s", recorder.Code, testCase.status, recorder.Body.String())
-			}
-			var response factoryapi.ErrorResponse
-			if err := json.Unmarshal(recorder.Body.Bytes(), &response); err != nil {
-				t.Fatalf("decode error response: %v", err)
-			}
-			if response.Code != testCase.code {
-				t.Fatalf("error code = %q, want %q", response.Code, testCase.code)
-			}
-		})
-	}
-}
-
-func TestGetWorkerSessionObservationBySessionIDRejectsUnsupportedIdentity(t *testing.T) {
-	service := &fakeObservationService{}
-	handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-	for _, testCase := range []struct {
-		name     string
-		provider string
-		kind     string
-		code     factoryapi.ErrorResponseCode
-	}{
-		{name: "provider", provider: "other", kind: "session_id", code: factoryapi.ErrorResponseCode("PROVIDER_UNSUPPORTED")},
-		{name: "kind", provider: "codex", kind: "rollout", code: factoryapi.ErrorResponseCode("SESSION_KIND_UNSUPPORTED")},
-	} {
-		t.Run(testCase.name, func(t *testing.T) {
-			recorder := httptest.NewRecorder()
-			request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions/detail", nil)
-			handler.GetWorkerSessionObservationBySessionId(recorder, request, factoryapi.SessionID("session-1"), factoryapi.GetWorkerSessionObservationBySessionIdParams{
-				Provider: factoryapi.LoadableProviderSessionProvider(testCase.provider), Kind: factoryapi.LoadableProviderSessionKind(testCase.kind), Id: "session-1",
-			})
-			if recorder.Code != http.StatusBadRequest {
-				t.Fatalf("status = %d, want 400", recorder.Code)
-			}
-			var response factoryapi.ErrorResponse
-			if err := json.Unmarshal(recorder.Body.Bytes(), &response); err != nil {
-				t.Fatalf("decode error response: %v", err)
-			}
-			if response.Code != testCase.code {
-				t.Fatalf("error code = %q, want %q", response.Code, testCase.code)
-			}
-			if service.getCalled {
-				t.Fatal("observation service called for unsupported identity")
-			}
-		})
-	}
-}
-
-func TestStreamWorkerSessionEventsBySessionIDWritesRetainedAndTerminalFrames(t *testing.T) {
-	service := &fakeObservationService{
-		getResult: workersessions.Observation{
-			WorkerSessionID: "worker-session-1", ProviderSessionAvailable: true,
-			ProviderSession: providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-1"},
-			WorkIDs:         []string{"work-1"}, State: workersessions.StateRunning,
-		},
-		streamSubscription: &fakeObservationSubscription{deliveries: []workersessions.ObservationDelivery{
-			{Kind: workersessions.ObservationDeliveryRecord, Event: workersessions.ObservationEvent{
-				Position: 1, SourceType: "worker_session", SourceID: "worker-session-1", SourceSequence: 1,
-				SourceEventID: "event-1", SchemaID: "worker_session.started", Payload: json.RawMessage(`{"state":"RUNNING"}`),
-			}},
-			{Kind: workersessions.ObservationDeliveryTerminal, Event: workersessions.ObservationEvent{
-				Position: 2, SourceType: "worker_session", SourceID: "worker-session-1", SourceSequence: 2,
-				SourceEventID: "event-2", SchemaID: "worker_session.completed", Payload: json.RawMessage(`{"state":"COMPLETED"}`),
-			}},
-		}},
-	}
-	handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-	recorder := httptest.NewRecorder()
-	request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions/events?provider=codex&kind=session_id&id=provider-session-1", nil)
-
-	handler.StreamWorkerSessionEventsBySessionId(recorder, request, factoryapi.SessionID("session-1"), factoryapi.StreamWorkerSessionEventsBySessionIdParams{
-		Provider: factoryapi.LoadableProviderSessionProvider("codex"), Kind: factoryapi.LoadableProviderSessionKind("session_id"), Id: "provider-session-1",
-	})
-	assertRetainedTerminalFrames(t, recorder, service)
-}
-
-func TestStreamWorkerSessionEventsBySessionIDReplayOnlyWritesSummaryAndPreservesMode(t *testing.T) {
-	replayOnly := true
-	service := &fakeObservationService{
-		getResult: workersessions.Observation{
-			WorkerSessionID: "worker-session-1", ProviderSessionAvailable: true,
-			ProviderSession: providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-1"},
-			WorkIDs:         []string{"work-1"}, State: workersessions.StateRunning,
-		},
-		streamSubscription: &fakeObservationSubscription{deliveries: []workersessions.ObservationDelivery{
-			{Kind: workersessions.ObservationDeliveryRecord, Event: workersessions.ObservationEvent{Position: 1, Payload: json.RawMessage(`{"step":1}`)}},
-			{Kind: workersessions.ObservationDeliveryReplaySummary, Summary: &workersessions.ReplaySummary{Complete: false, Reason: "session-active", EventsEmitted: 1}},
-		}},
-	}
-	handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-	recorder := httptest.NewRecorder()
-	request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions/events?provider=codex&kind=session_id&id=provider-session-1&replayOnly=true", nil)
-
-	handler.StreamWorkerSessionEventsBySessionId(recorder, request, factoryapi.SessionID("session-1"), factoryapi.StreamWorkerSessionEventsBySessionIdParams{
-		Provider: factoryapi.LoadableProviderSessionProvider("codex"), Kind: factoryapi.LoadableProviderSessionKind("session_id"), Id: "provider-session-1", ReplayOnly: &replayOnly,
-	})
-
-	frames := decodeSSEFrames(t, recorder.Body.String())
-	if len(frames) != 2 || frames[0].Delivery != "RECORD" || frames[1].Delivery != "REPLAY_SUMMARY" {
-		t.Fatalf("frames = %#v, want RECORD then REPLAY_SUMMARY", frames)
-	}
-	if frames[1].ReplaySummary == nil || frames[1].ReplaySummary.Kind != "replay-summary" || frames[1].ReplaySummary.Complete || frames[1].ReplaySummary.Reason != "session-active" || frames[1].ReplaySummary.EventsEmitted != 1 {
-		t.Fatalf("replay summary = %#v, want active count-one summary", frames[1].ReplaySummary)
-	}
-	if !service.streamRequest.ReplayOnly {
-		t.Fatal("adapter did not preserve replay-only mode in the Worker Sessions request")
-	}
-}
-
-func TestStreamWorkerSessionEventsBySessionIDWritesExplicitSourceFailure(t *testing.T) {
-	service := &fakeObservationService{
-		getResult: workersessions.Observation{
-			WorkerSessionID: "worker-session-1", ProviderSessionAvailable: true,
-			ProviderSession: providers.SessionRef{Provider: providers.IDCursor, Kind: providers.SessionIDKind, ID: "cursor-session-1"},
-		},
-		streamSubscription: &fakeObservationSubscription{deliveries: []workersessions.ObservationDelivery{
-			{Kind: workersessions.ObservationDeliverySourceFailure, Err: workersessions.ErrObservationSourceGap},
-		}},
-	}
-	handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-	recorder := httptest.NewRecorder()
-	request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions/events", nil)
-
-	handler.StreamWorkerSessionEventsBySessionId(recorder, request, factoryapi.SessionID("session-1"), factoryapi.StreamWorkerSessionEventsBySessionIdParams{
-		Provider: factoryapi.LoadableProviderSessionProvider("cursor"), Kind: factoryapi.LoadableProviderSessionKind("session_id"), Id: "cursor-session-1",
-	})
-
-	frames := decodeSSEFrames(t, recorder.Body.String())
-	if len(frames) != 1 || frames[0].Delivery != "SOURCE_FAILURE" || frames[0].Event != nil {
-		t.Fatalf("frames = %#v, want one source failure without an event", frames)
-	}
-	if frames[0].ErrorCode == nil || *frames[0].ErrorCode != "WORKER_SESSION_STREAM_GAP" {
-		t.Fatalf("error code = %#v, want WORKER_SESSION_STREAM_GAP", frames[0].ErrorCode)
-	}
-	if frames[0].ErrorMessage == nil || !strings.Contains(*frames[0].ErrorMessage, "retained") {
-		t.Fatalf("error message = %#v, want safe retained-history message", frames[0].ErrorMessage)
-	}
-}
-
-func TestStreamWorkerSessionEventsBySessionIDMapsUnavailableBeforeOpening(t *testing.T) {
-	service := &fakeObservationService{getResult: workersessions.Observation{
-		WorkerSessionID: "worker-session-1", ProviderSessionAvailable: true,
-		ProviderSession: providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-1"},
-	}, streamErr: workersessions.ErrObservationSourceUnavailable}
-	handler := NewHandler(NewAdapter(service, workServiceStub{}), zap.NewNop())
-	recorder := httptest.NewRecorder()
-	request := httptest.NewRequest("GET", "/factory-sessions/session-1/worker-sessions/events", nil)
-
-	handler.StreamWorkerSessionEventsBySessionId(recorder, request, factoryapi.SessionID("session-1"), factoryapi.StreamWorkerSessionEventsBySessionIdParams{
-		Provider: factoryapi.LoadableProviderSessionProvider("codex"), Kind: factoryapi.LoadableProviderSessionKind("session_id"), Id: "provider-session-1",
-	})
-
-	if recorder.Code != http.StatusInternalServerError {
-		t.Fatalf("status = %d, want 500; body=%s", recorder.Code, recorder.Body.String())
-	}
-	var response factoryapi.ErrorResponse
-	if err := json.Unmarshal(recorder.Body.Bytes(), &response); err != nil {
-		t.Fatalf("decode error response: %v", err)
-	}
-	if response.Code != factoryapi.ErrorResponseCodeWORKERSESSIONSTREAMUNAVAILABLE {
-		t.Fatalf("error code = %q, want WORKER_SESSION_STREAM_UNAVAILABLE", response.Code)
 	}
 }
